refactor(gbd): convert to float before round

+++ ihme_gbd/gbd_tools.py
@@ -265,7 +265,7 @@ def create_variables(
 def clean_units_and_values(df: pd.DataFrame) -> pd.DataFrame:
 
     df["val"][df["metric"] == "Percent"] = df["val"][df["metric"] == "Percent"] * 100
-    df["val"] = df["val"].round(2)
+    df["val"] = df["val"].astype(float).round(2)
     df["val"][
         (df["measure"].isin(["Prevalence", "Incidence", "Deaths"]))
         & (df["metric"] == "Number")
